Refactor(core): Improve listener architecture and thread safety
Refactor(core): Improve listener architecture and thread safety

This commit refactors the event listener mechanism within the core
blockchain processing components to improve modularity, thread safety,
and overall code quality.

The previous implementation had several architectural issues:
- Listener interfaces were defined as inner classes within
  `BlockchainProcessorImpl`, violating the Single Responsibility
  Principle (SRP) and creating tight coupling with consumers like the GUI.
- Listener lists used `java.util.ArrayList`, which is not thread-safe
  and could lead to `ConcurrentModificationException` in a
  multi-threaded environment.
- The public API defined in the `BlockchainProcessor` interface did not
  fully represent the component's contract (e.g., it was missing
  listener management and shutdown methods).

This refactoring addresses these points by:
1.  **Extracting Listener Interfaces:** All listener interfaces
    (`PeerCountListener`, `PerformanceListener`, etc.) are now in their
    own files within a dedicated `brs.listeners` package. This improves
    modularity and makes the API clearer.

2.  **Enhancing Thread Safety:** All listener collections in
    `BlockchainProcessorImpl` have been changed from `ArrayList` to
    `CopyOnWriteArrayList`. This ensures safe iteration during event
    dispatching without synchronization overhead on reads.

3.  **Defining a Clear Public Contract:**
    - Listener management methods (`add...Listener`, `remove...Listener`)
      and the `shutdown()` method have been added to the
      `BlockchainProcessor` interface.
    - `Signum.getBlockchainProcessor()` now returns the `BlockchainProcessor`
      interface instead of the concrete implementation, promoting loose
      coupling and programming to an interface.
