+++ src/brs/BlockchainProcessor.java
@@ -1,5 +1,9 @@
 package brs;
 
+import brs.listeners.NetVolumeListener;
+import brs.listeners.PeerCountListener;
+import brs.listeners.PerformanceListener;
+import brs.listeners.QueueStatusListener;
 import brs.peer.Peer;
 import brs.util.JSON;
 import brs.util.Observable;
@@ -32,6 +36,24 @@ public interface BlockchainProcessor extends Observable<Block, BlockchainProcess
 
     void generateBlock(String secretPhrase, byte[] publicKey, Long nonce)
             throws BlockNotAcceptedException;
+
+    void shutdown();
+
+    void addQueueStatusListener(QueueStatusListener listener);
+
+    void removeQueueStatusListener(QueueStatusListener listener);
+
+    void addPerformanceListener(PerformanceListener listener);
+
+    void removePerformanceListener(PerformanceListener listener);
+
+    void addPeerCountListener(PeerCountListener listener);
+
+    void removePeerCountListener(PeerCountListener listener);
+
+    void addNetVolumeListener(NetVolumeListener listener);
+
+    void removeNetVolumeListener(NetVolumeListener listener);
 
     List<Block> popOffTo(int height);
