+++ src/brs/listeners/NetVolumeListener.java
@@ -1,0 +1,5 @@
+package brs.listeners;
+
+public interface NetVolumeListener {
+    void onNetVolumeChanged(long uploadedVolume, long downloadedVolume);
+}
